[fix/011] 값 변경 로직 수정
