+++ src/main/java/com/s3k/backend/weather/api_client/kma/entity/KmaFeignResponse.java
@@ -1,5 +1,6 @@
 package com.s3k.backend.weather.api_client.kma.entity;
 
+import com.s3k.backend.weather.entity.Weather;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -11,9 +12,8 @@ public class KmaFeignResponse {
   private Header header;
   private Body body;
 
-  // 만든 리스트에서 현재시간에 맞춰서 다음 시간부터?
-  // 아니면 현재시간을 기준으로?
-  public List<Item> getItems() {
-    return body.getItems().getItem();
+  public List<Weather> getWeatherList() {
+    return body.getWeatherList();
   }
+
 }
