+++ src/main/java/com/s3k/backend/weather/dto/WeatherDto.java
@@ -1,0 +1,14 @@
+package com.s3k.backend.weather.dto;
+
+import com.s3k.backend.weather.entity.Weather;
+import java.util.List;
+
+public class WeatherDto {
+
+  public record Response(
+      List<Weather> weathers
+  ) {
+
+  }
+
+}
